requirements.txt
# ============================================================================
# FILE 1: requirements.txt
# ============================================================================

fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python
numpy
onnxruntime
pydantic>=2.0.0
pydantic-settings
python-dotenv
aiofiles
httpx
loguru
pytest
pytest-asyncio